Move iterative solvers to own module, implement get_ and set_options
